feat: Rework the api for proper OpenAPI
Rework the federation and passkey api to ensure we can generate the
CLI/SDK code from the resulting OpenAPI.
